aut/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.18.1
pika~=1.3.2